Added custom hybridpersistent_filesystem attribute
By default kiwi creates a btrfs partition as write space for hybrid
live systems. With the new attribute one can specify also the 'fat'
filesystem as write space. This would allow to create hybrid live
images which are compatible with the Windows OS in a similar way
like kiwi's livestick tool does it. This Fixes #475