Added type attribute wwid_wait_timeout
multipath devices from a SAN system take some time to appear on
the host. In order to have all presented wwn's available on the
host it's required to wait some time because we can't poll for
the appearance of a specific device the wait period is now
customizable. The default wait period is 3 seconds