Support for stacked zipl plus grub2 boot
In SLES12 zipl and btrfs are incompatible because the block where
the initrd and the kernel live could change in a btrfs filesystem.
Because IBM is not going to fix zipl to work with btrfs a new boot
process has been invented by the SLES team which impacts kiwi too.