Added a force option to package deletion
zypper is not really capable to remove a package without
being intelligent. There is no way to pass rpm options like
nodeps, noscripts or similar to the call. Thus if force
is activated for the zypper package manager instance we use
the plain rpm call to just remove the thing